tests: move getSomeWdEditionUri to entities fixtures
