+++ tests/api/entities/update_wd_claims.test.ts
@@ -1,28 +1,8 @@
-import { random } from 'lodash-es'
 import should from 'should'
-import wdk from 'wikibase-sdk/wikidata.org'
-import { prefixifyWd } from '#controllers/entities/lib/prefix'
-import { someRandomImageHash } from '#fixtures/entities'
+import { getSomeWdEditionUri, someRandomImageHash } from '#fixtures/entities'
 import { validateP31Update } from '#lib/wikidata/validate_wd_update'
-import type { Url } from '#server/types/common'
-import type { WdEntityId } from '#server/types/entity'
 import { addClaim, getByUri, removeClaim, updateClaim } from '#tests/api/utils/entities'
-import { request } from '#tests/api/utils/request'
 import { shouldNotBeCalled } from '#tests/unit/utils/utils'
-
-const { cirrusSearchPages, parse } = wdk
-
-async function getSomeWdEditionUri () {
-  const url = cirrusSearchPages({
-    haswbstatement: [ 'P31=Q3331189', 'P629' ],
-    limit: 1,
-    offset: random(0, 10000),
-    prop: [],
-  }) as Url
-  const res = await request('get', url)
-  const id = parse.pagesTitles(res)[0] as WdEntityId
-  return prefixifyWd(id)
-}
 
 describe('entities:update-claims:wd', () => {
   describe('validateP31Update', () => {
